+++ ext/tableFilter.js
@@ -2891,12 +2891,13 @@ layui.define(['table', 'form', 'laydate', 'util', 'excel', 'laytpl'], function (
                             }
                             return val;
                         }
+                        var value = bodyIndex >= 0 && columnsMap[columnsMap.length - 1][field].templet ?
+                            typeof columnsMap[columnsMap.length - 1][field].templet === 'function' ?
+                                $('<div>' + columnsMap[columnsMap.length - 1][field].templet(line) + '</div>').find(':input').length===0?$('<div>' + columnsMap[columnsMap.length - 1][field].templet(line) + '</div>').text():$tableBody.children('tbody').children('tr[data-index='+bodyIndex+']').children('td[data-field="'+field+'"]').find(':input').val() || handleNull(line[field])
+                                : $('<div>'+laytpl($(columnsMap[columnsMap.length - 1][field].templet).html() || String(columnsMap[columnsMap.length - 1][field].templet)).render(line)+'</div>').find(':input').length===0?$('<div>'+laytpl($(columnsMap[columnsMap.length - 1][field].templet).html() || String(columnsMap[columnsMap.length - 1][field].templet)).render(line)+'</div>').text():$tableBody.children('tbody').children('tr[data-index='+bodyIndex+']').children('td[data-field="'+field+'"]').find(':input').val() || handleNull(line[field])
+                            : bodyIndex >=0 && columnsMap[columnsMap.length - 1][field].type === 'numbers' ? bodyIndex+1 : handleNull(line[field]);
                         return {
-                            v: bodyIndex >= 0 && columnsMap[columnsMap.length - 1][field].templet ?
-                                typeof columnsMap[columnsMap.length - 1][field].templet === 'function' ?
-                                    $('<div>' + columnsMap[columnsMap.length - 1][field].templet(line) + '</div>').find(':input').length===0?$('<div>' + columnsMap[columnsMap.length - 1][field].templet(line) + '</div>').text():$tableBody.children('tbody').children('tr[data-index='+bodyIndex+']').children('td[data-field="'+field+'"]').find(':input').val() || handleNull(line[field])
-                                    : $('<div>'+laytpl($(columnsMap[columnsMap.length - 1][field].templet).html() || String(columnsMap[columnsMap.length - 1][field].templet)).render(line)+'</div>').find(':input').length===0?$('<div>'+laytpl($(columnsMap[columnsMap.length - 1][field].templet).html() || String(columnsMap[columnsMap.length - 1][field].templet)).render(line)+'</div>').text():$tableBody.children('tbody').children('tr[data-index='+bodyIndex+']').children('td[data-field="'+field+'"]').find(':input').val() || handleNull(line[field])
-                                : bodyIndex >=0 && columnsMap[columnsMap.length - 1][field].type === 'numbers' ? bodyIndex+1 : handleNull(line[field]),// v 代表单元格的值
+                            v: value,// v 代表单元格的值
                             s: {// s 代表样式
                                 alignment: {
                                     horizontal: columnsMap[bodyIndex<-1 ? curIndex - tableStartIndex + 1 : columnsMap.length - 1][field].align ? alignTrans[columnsMap[bodyIndex<-1 ? curIndex - tableStartIndex + 1 : columnsMap.length - 1][field].align] : 'top',
@@ -2908,7 +2909,7 @@ layui.define(['table', 'form', 'laydate', 'util', 'excel', 'laytpl'], function (
                                 },
                                 border: border
                             },
-                            t: cellType
+                            t: UNHANDLED_VALUES.indexOf(value) === -1 ? cellType : 's'
                         };
                     }
                 }